+++ scripts/deploy-canonical.js
@@ -2,6 +2,7 @@ require("dotenv").config();
 const hre = require("hardhat");
 const ethers = hre.ethers;
 const path = require("path");
+const fs = require("fs-extra");
 const EthDeployUtils = require("eth-deploy-utils");
 const bytecodes = require("../test/helpers/bytecodes.json");
 let deployUtils;
@@ -31,6 +32,7 @@ async function main() {
       "0x0000000000000000000000000000000000000000fd8eb4e1dca713016c518e31",
     );
     await deployUtils.deployContractViaNickSFactory(deployer, "CrunaRegistry", undefined, undefined, salt);
+
     await deployUtils.deployContractViaNickSFactory(
       deployer,
       "CrunaGuardian",
@@ -40,7 +42,34 @@ async function main() {
     );
   } else {
     await deployUtils.deployBytecodeViaNickSFactory(deployer, "CrunaRegistry", canonicalBytecodes.CrunaRegistry);
-    await deployUtils.deployBytecodeViaNickSFactory(deployer, "CrunaGuardian", canonicalBytecodes.CrunaGuardian);
+
+    if (process.env.RECODE_GUARDIAN) {
+      // This is supposed to happen only during development when there are breaking changes.
+      // It should not happen after the first guardian as been deployed, except if very serious
+      // issues are found and a new guardian is needed. In that unfortunate case, all managers
+      // and plugins will have to be upgraded by tokens' owners.
+      canonicalBytecodes.CrunaGuardian = await deployUtils.getBytecodeToBeDeployedViaNickSFactory(
+        deployer,
+        "CrunaGuardian",
+        ["uint256", "address[]", "address[]", "address"],
+        [delay, [proposerAddress], [executorAddress], deployer.address],
+        salt,
+      );
+      const guardian = await deployUtils.deployBytecodeViaNickSFactory(
+        deployer,
+        "CrunaGuardian",
+        canonicalBytecodes.CrunaGuardian,
+      );
+      fs.writeFileSync(
+        path.resolve(__dirname, "../contracts/canonicalBytecodes.json"),
+        JSON.stringify(canonicalBytecodes, null, 2),
+      );
+      let canonical = fs.readFileSync(path.resolve(__dirname, "../libs-canonical/not-localhost/Canonical.sol"), "utf8");
+      canonical = canonical.replace(/ICrunaGuardian\([^)]+\)/, `ICrunaGuardian(${guardian.address})`);
+      fs.writeFileSync(path.resolve(__dirname, "../libs-canonical/not-localhost/Canonical.sol"), canonical);
+    } else {
+      await deployUtils.deployBytecodeViaNickSFactory(deployer, "CrunaGuardian", canonicalBytecodes.CrunaGuardian);
+    }
   }
 }